+++ config/astrotomic-laravel-translatable-plugin.php
@@ -1,0 +1,6 @@
+<?php
+
+// config for CorneelD/LaravelTranslatablePlugin
+return [
+
+];