+++ src/AstrotomicLaravelTranslatable.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace CorneelD\AstrotomicLaravelTranslatablePlugin;
+
+class AstrotomicLaravelTranslatable
+{
+}