+++ src/FilamentLaravelLocalizationPlugin.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace CorneelD\FilamentLaravelLocalizationPlugin;
-
-class FilamentLaravelLocalizationPlugin
-{
-}