+++ src/Testing/TestsAstrotomicLaravelTranslatablePlugin.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace CorneelD\AstrotomicLaravelTranslatablePlugin\Testing;
+
+use Livewire\Features\SupportTesting\Testable;
+
+/**
+ * @mixin Testable
+ */
+class TestsAstrotomicLaravelTranslatablePlugin
+{
+    //
+}